# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Platform Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Agent {
  id: string;
  name: string;
  category: string;
  description: string;
  color?: string;
  emoji?: string;
  vibe?: string;
  systemPrompt: string;
  filePath: string;
  isActive: boolean;
  createdAt: string;
}

export interface AgentListResponse {
  agents: Agent[];
  total: number;
  categories: string[];
}

export interface ReloadAgentsResponse {
  loaded: number;
  message: string;
}

export interface Project {
  id: string;
  name: string;
  description?: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectListResponse {
  projects: Project[];
}

export interface CreateProjectRequest {
  name: string;
  description?: string;
}

export interface Task {
  id: string;
  projectId: string;
  agentId?: string;
  title: string;
  description?: string;
  status: string;
  result?: string;
  model?: string;
  ollamaEndpoint?: string;
  createdAt: string;
  updatedAt: string;
}

export interface TaskListResponse {
  tasks: Task[];
}

export interface UpdateProjectRequest {
  name?: string;
  description?: string;
}

export interface CreateTaskRequest {
  projectId: string;
  agentId?: string;
  title: string;
  description?: string;
  model?: string;
  ollamaEndpoint?: string;
}

export interface UpdateTaskRequest {
  title?: string;
  model?: string;
}

export interface RunTaskRequest {
  message: string;
  model?: string;
  ollamaEndpoint?: string;
  enableThinking?: boolean;
}

export interface Message {
  id: string;
  taskId: string;
  agentId?: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface MessageListResponse {
  messages: Message[];
}

export interface ContextEntry {
  id: string;
  projectId: string;
  key: string;
  value: string;
  source?: string;
  createdAt: string;
}

export interface ContextListResponse {
  context: ContextEntry[];
}

export interface SuggestAgentRequest {
  task: string;
  projectId?: string;
}

export interface SuggestAgentResponse {
  agent: Agent;
  reason: string;
  alternatives: Agent[];
}

export interface PlanProjectRequest {
  prompt: string;
  model: string;
}

export interface PlanTask {
  title: string;
  agentId: string;
  agentName: string;
  prompt: string;
  reason: string;
}

export interface PlanProjectResponse {
  tasks: PlanTask[];
}

export interface OllamaModel {
  name: string;
  size: number;
  modifiedAt: string;
}

export interface OllamaModelsResponse {
  models: OllamaModel[];
  endpoint: string;
}

export interface OllamaStatusResponse {
  connected: boolean;
  endpoint: string;
  version?: string;
  message?: string;
}

export type OllamaChatRequestMessagesItem = {
  role: string;
  content: string;
};

export interface OllamaChatRequest {
  model: string;
  messages: OllamaChatRequestMessagesItem[];
  systemPrompt?: string;
  enableThinking?: boolean;
  ollamaEndpoint?: string;
}

export type ListAgentsParams = {
  category?: string;
  search?: string;
};

export type ListOllamaModelsParams = {
  endpoint?: string;
};

export type GetOllamaStatusParams = {
  endpoint?: string;
};