# lib/api-zod/src/generated/types/ollamaChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Platform Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { OllamaChatRequestMessagesItem } from "./ollamaChatRequestMessagesItem";

export interface OllamaChatRequest {
  model: string;
  messages: OllamaChatRequestMessagesItem[];
  systemPrompt?: string;
  enableThinking?: boolean;
  ollamaEndpoint?: string;
}