# lib/api-zod/src/generated/types/ollamaModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Platform Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface OllamaModel {
  name: string;
  size: number;
  modifiedAt: string;
}