# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Platform Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: string;
  projectId: string;
  agentId?: string;
  title: string;
  description?: string;
  status: string;
  result?: string;
  model?: string;
  ollamaEndpoint?: string;
  createdAt: string;
  updatedAt: string;
}