# lib/api-zod/src/generated/types/updateProjectRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Platform Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectRequest {
  name?: string;
  description?: string;
}